Add files via upload
Piezo Electrical Tranducers Simulation in COMSOL
Diaphragm simulation : diaphragm is made of a base with a radius of 100um and made of polycrystalline silicon. A PZT-2 layer with a radius of 50um is placed on it as a Piezo element. Finally, the amount of changes in the presence of an electric field with a potential of 500v has been determined. Obviously, in this simulation that shows the performance of the piezoelectric transducer, electrical energy is converted into mechanical energy. In other words, the location of the diaphragm can change depending on the electrical energy. 
Also the simulated sensor consists of a 200um base with a diameter of 5um, which is made of polycrystalline silicon, and a PZT-2 layer with a length of 75um and a diameter of 5um. Both layers are coupled together and the simulation shows the mechanical changes of the sensor by applying 0.01 Newton from the corner of the base layer. Explanation: The sensor converts mechanical changes into electrical voltage